+++ contracts/common/IBaseHTS.sol
@@ -16,7 +16,7 @@ interface IBaseHTS {
         address token,
         address sender,
         address receiver,
-        int256 serial
+        int64 serial
     ) external returns (int256);
 
     function associateTokenPublic(
