+++ contracts/mock/MockBaseHTSWithTokenCreationFail.sol
@@ -58,6 +58,6 @@ contract MockBaseHTSWithTokenCreationFail is IBaseHTS {
         address token,
         address sender,
         address receiver,
-        int256 serial
+        int64 serial
     ) external override returns (int256) {}
 }
